% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Warprrr.R
\name{get_data}
\alias{get_data}
\title{Generic Data Getter for warprrr Class}
\usage{
get_data(warper, verbose, ...)
}
\arguments{
\item{verbose}{Print detailed messages if TRUE.}

\item{warprrr}{warprrr object.}
}
\value{
Data.table, tibble, or arrow table, as appropriate.
}
\description{
Loads data from cache if available, otherwise reads source and caches result.
}
\examples{
# get_data(my_warprrr_object)
}
