% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inform_glue}
\alias{inform_glue}
\title{Print a formatted, blue-colored glue message}
\usage{
inform_glue(..., envir = parent.frame())
}
\arguments{
\item{...}{Arguments passed to \code{glue::glue()} for formatting.}

\item{envir}{Environment passed to \code{glue::glue}.
Default is \code{parent.frame()}.}
}
\description{
Uses \code{glue} for string interpolation and \code{cli} for colored output.
}
