% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_path}
\alias{pkg_path}
\title{Generate a normalized package resource file path.}
\usage{
pkg_path(..., allow_error = FALSE)
}
\arguments{
\item{...}{Character vectors specifying file path components within the package.}

\item{allow_error}{Logical. If \code{TRUE}, throw an error if the file is missing; otherwise, return \code{fs::path("")}.}
}
\value{
An \code{fs_path} object (normalized file path) or an empty path if missing and \code{allow_error = FALSE}.
}
\description{
Returns the normalized file path for a resource in the \code{cranpkgtemplate} package.
If the file does not exist and \code{allow_error = TRUE}, an error is thrown. If \code{allow_error = FALSE},
returns an empty \code{fs_path} object.
}
