% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_taken_precise}
\alias{time_taken_precise}
\title{Measure Precise Evaluation Time}
\usage{
time_taken_precise(expr)
}
\arguments{
\item{expr}{Expression to be evaluated.}
}
\value{
Numeric value: elapsed time in seconds.
}
\description{
Returns elapsed time in seconds (to six decimals) for the given expression.
}
