% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Warprrr.R
\name{inform_glue_verbose}
\alias{inform_glue_verbose}
\title{Verbose Informational Message with glue}
\usage{
inform_glue_verbose(..., verbose, envir = parent.frame())
}
\arguments{
\item{...}{Message arguments for glue.}

\item{verbose}{Logical; print message if TRUE.}

\item{envir}{The parent env when func is called. Required for glue to work.}
}
\value{
Invisible NULL.
}
\description{
Optionally prints a glue message if verbose is TRUE.
}
