% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Title}
\usage{
read_data(data_path, verbose = FALSE, ...)
}
\arguments{
\item{data_path}{The path to data. csv, psv, tsv, parquet, feather, txt,
sas7bdat, xpt file formats are supported}

\item{...}{Arguments passed onto the respective read function. csv, psv, tsv,
txt, is read with data.table::fread. sas7bdat and xpt are read with
haven::read_sas and haven::read_xpt respectively. parquet and feather
are read with arrow::read_parquet and arrow::read_feather respectively.
Arguments can be passed to these functions via this argument.}
}
\value{
An object of class data.frame.
}
\description{
Title
}
