% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Title}
\usage{
read_data(
  data_path,
  cache_path = "~/.cache/patient_profiles/",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data_path}{The path to data. csv, psv, tsv, parquet, feather, txt,
sas7bdat, xpt file formats are supported}

\item{cache_path}{The path to use to store the cache.}

\item{verbose}{Whether to print logs, time taken to read non-cached vs
cached data etc.}

\item{...}{Arguments passed onto the respective read function. csv, psv, tsv,
txt, is read with data.table::fread. sas7bdat and xpt are read with
haven::read_sas and haven::read_xpt respectively. parquet and feather
are read with arrow::read_parquet and arrow::read_feather respectively.
Arguments can be passed to these functions via this argument.}
}
\value{
A named list of the form: list(data = data, warprrr = warprrr_class).
warprrr class can be used to observe or use the properties of this instance
of warprrr.
}
\description{
Title
}
