% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Warprrr.R
\name{warprrr}
\alias{warprrr}
\title{warprrr S7 Class for Data Wrangling and Caching}
\usage{
warprrr(
  data_path = character(0),
  read_fun_args = list(),
  cache_path = "~/.cache/patient_profiles/",
  cache_ext = ".feather"
)
}
\description{
Defines an S7 class for handled data loading and caching in Parquet format.
}
